requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --universal --output-file requirements.txt
absl-py==2.4.0
    # via
    #   keras
    #   tensorflow
annotated-types==0.7.0
    # via pydantic
astunparse==1.6.3
    # via tensorflow
certifi==2026.2.25
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.5
    # via requests
click==8.3.1
    # via wandb
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
flatbuffers==25.12.19
    # via tensorflow
fonttools==4.62.0
    # via matplotlib
gast==0.7.0
    # via tensorflow
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via wandb
google-pasta==0.2.0
    # via tensorflow
grpcio==1.78.0
    # via tensorflow
h5py==3.14.0
    # via
    #   keras
    #   tensorflow
idna==3.11
    # via requests
joblib==1.5.3
    # via scikit-learn
keras==3.13.2
    # via tensorflow
kiwisolver==1.5.0
    # via matplotlib
libclang==18.1.1
    # via tensorflow
markdown-it-py==4.0.0
    # via rich
matplotlib==3.10.8
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.4
    # via
    #   keras
    #   tensorflow
namex==0.1.0
    # via keras
numpy==2.4.3
    # via
    #   -r requirements.in
    #   contourpy
    #   h5py
    #   keras
    #   matplotlib
    #   ml-dtypes
    #   opencv-python
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorflow
opencv-python==4.13.0.92
    # via -r requirements.in
opt-einsum==3.4.0
    # via tensorflow
optree==0.19.0
    # via keras
packaging==26.0
    # via
    #   keras
    #   matplotlib
    #   tensorflow
    #   wandb
    #   wheel
pandas==3.0.1
    # via -r requirements.in
pillow==12.1.1
    # via matplotlib
platformdirs==4.9.4
    # via wandb
protobuf==6.33.5
    # via
    #   tensorflow
    #   wandb
pydantic==2.12.5
    # via wandb
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pyyaml==6.0.3
    # via wandb
requests==2.32.5
    # via
    #   tensorflow
    #   wandb
rich==14.3.3
    # via keras
scikit-learn==1.8.0
    # via -r requirements.in
scipy==1.17.1
    # via scikit-learn
sentry-sdk==2.54.0
    # via wandb
setuptools==82.0.1
    # via tensorflow
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorflow
smmap==5.0.3
    # via gitdb
tensorflow==2.21.0
    # via -r requirements.in
termcolor==3.3.0
    # via tensorflow
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   grpcio
    #   optree
    #   pydantic
    #   pydantic-core
    #   tensorflow
    #   typing-inspection
    #   wandb
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3 ; sys_platform == 'emscripten' or sys_platform == 'win32'
    # via pandas
urllib3==2.6.3
    # via
    #   requests
    #   sentry-sdk
wandb==0.25.1
    # via -r requirements.in
wheel==0.46.3
    # via astunparse
wrapt==2.1.2
    # via tensorflow
